+++ sdk/spring/azure-spring-cloud-autoconfigure/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.7.0-beta.1 (Unreleased)
 ### Key Bug Fixes
 - Fixed `EventHubMessageConverter` to load all system properties of `EventData` and put in the header of org.springframework.messaging.Message.([#22683](https://github.com/Azure/azure-sdk-for-java/pull/22683/))
+- Fix bug of setting message headers repeatedly with different value types. ([#22939](https://github.com/Azure/azure-sdk-for-java/pull/22939))
 
 
 ## 2.6.0 (2021-06-23)